packages: build GmsCompat by default
* I don't understand the aspects behind it all but with devices that
use the Generic System Image standard practices (Google Pixel 6 Pro),
then there are tight restrictions on artifact path requirements. You
overcome this by adding the offending files to an "allowlist" to
supress the error. But for some reason when an app that has files that
violate the requirements and need to add it files to the allowlist and
calling the compile of the offending package from the "build" repo the
artifact paths allowlist is not honored and the compile fails just move
the call to build GmsCompat to vendor/spark. It has no regressions and
makes more logical sense

Signed-off-by: Matt Filetto <matt.filetto@gmail.com>